updated: unit tests migration almost complete
