+++ test/specs/browser/_compiler.js
@@ -33,300 +33,6 @@ describe('Compiler Browser', function() {
 
 
 
-
-
-
-
-
-  it('style injection to single style tag', function() {
-    var styles = getRiotStyles()
-
-    expect(styles).to.match(/\bp\s*\{color: blue;}/)
-    expect(styles).to.match(/\bdiv\s*\{color: red;}/)
-  })
-
-  // working
-  it('style injection removes type riot style tag', function() {
-    var stag = document.querySelector('style[type=riot]')
-    expect(stag).to.be.equal(null)
-  })
-
-  if (typeof window.__karma__ === 'undefined') {
-    it('style tag sits in between title and link to stylesheet', function () {
-      var stag = document.querySelector('style')
-      var prevE = stag.previousElementSibling
-      var nextE = stag.nextElementSibling
-      expect(prevE.tagName).to.be.equal('TITLE')
-      expect(nextE.tagName).to.be.equal('LINK')
-    })
-  }
-
-  it('scoped css tag supports htm5 syntax, multiple style tags', function () {
-
-    checkCSS(riot.mount('style-tag3')[0], '4px')
-    checkCSS(riot.mount('style-tag4')[0], '2px', 1)
-    delete riot.parsers.css.cssup
-
-    function checkCSS(t, x, p2) {
-      t.update()
-      var e = t.root.firstElementChild
-      expect(e.tagName).to.be.equal('P')
-      expect(window.getComputedStyle(e, null).borderTopWidth).to.be.equal(x)
-      if (p2) {
-        e = t.root.getElementsByTagName('P')[1]
-        expect(e.innerHTML).to.be.equal('x')
-        expect(window.getComputedStyle(e, null).borderTopWidth).to.be.equal('1px')
-      }
-      tags.push(t)
-    }
-  })
-
-  it('scoped css and data-is, mount(selector, tagname)', function() {
-
-
-    function checkBorder(t) {
-      var e = t.root.firstElementChild
-      var s = window.getComputedStyle(e, null).borderTopWidth
-      expect(s).to.be.equal('1px')
-
-    }
-
-    injectHTML([
-      '<scoped-tag></scoped-tag>',
-      '<div data-is="scoped-tag"></div>',
-      '<div id="scopedtag"></div>'
-    ])
-
-    var stags = riot.mount('scoped-tag')
-
-    var tag = stags[0]
-    checkBorder(tag)
-
-    var rtag = stags[1]
-    checkBorder(rtag)
-
-    var divtag = riot.mount('#scopedtag', 'scoped-tag')[0]
-    checkBorder(divtag)
-    tags.push(divtag)
-    tags.push(rtag)
-    tag.unmount()
-  })
-
-  it('deferred injection of styles in batch', function() {
-
-    // test riot.util.styleNode
-    expect(riot.util.styleNode).to.not.be.equal(undefined)
-    expect(riot.util.styleNode.tagName).to.be.equal('STYLE')
-
-    // test style isn't injected yet
-    styles = getRiotStyles()
-    expect(styles).not.to.match(/\bparsed-style\s*\{/)
-
-    // define a styled tag
-    riot.tag('runtime-style-parsing', '<div></div>', '.parsed-style { color: red; }', '', function(opts) { })
-
-    // test style isn't injected by the simple tag definition
-    styles = getRiotStyles()
-    expect(styles).not.to.match(/\bparsed-style\s*\{/)
-
-    // mount the tag
-    injectHTML(['<runtime-style-parsing></runtime-style-parsing>' ])
-    var tag = riot.mount('runtime-style-parsing')[0]
-
-    // test style is correctly injected
-    styles = getRiotStyles()
-    expect(styles).to.match(/\bparsed-style\s*\{\s*color:\s*red;\s*}/)
-
-    // remount (unmount+mount)
-    tag.unmount(true)
-    tag = riot.mount('runtime-style-parsing')[0]
-    expect(tag).to.not.be.equal(undefined)
-
-    // test remount does not affect style
-    styles = getRiotStyles()
-    expect(styles).to.match(/\bparsed-style\s*\{\s*color:\s*red;\s*}/)
-
-    // test remount does not duplicate rule
-    expect(styles.match(/\bparsed-style\s*\{/g)).to.have.length(1)
-  })
-
-  it('preserve attributes from tag definition', function() {
-    injectHTML('<div data-is="preserve-attr2"></div>')
-    var tag = riot.mount('preserve-attr')[0]
-    expect(tag.root.className).to.be.equal('single-quote')
-    var tag2 = riot.mount('preserve-attr2')[0]
-    expect(tag2.root.className).to.be.equal('double-quote')
-    tag.unmount()
-    tags.push(tag2)
-  })
-
-  it('precompiled tag compatibility', function() {
-
-    injectHTML('<precompiled></precompiled>')
-    riot.tag('precompiled', 'HELLO!', 'precompiled, [data-is="precompiled"]  { color: red }', function(opts) {
-      this.nothing = opts.nothing
-    })
-
-    var tag = riot.mount('precompiled')[0]
-    expect(window.getComputedStyle(tag.root, null).color).to.be.equal('rgb(255, 0, 0)')
-    tag.unmount()
-
-  })
-
-  it('static named tag for tags property', function() {
-    injectHTML('<named-child-parent></named-child-parent>')
-    var tag = riot.mount('named-child-parent')[0]
-    expect(tag['tags-child'].root.innerHTML).to.be.equal('I have a name')
-
-    tag.unmount()
-  })
-
-  it('protect the internal "tags" attribute from external overrides', function() {
-    var tag = riot.mount('loop-protect-internal-attrs')[0]
-    expect(tag.tags['loop-protect-internal-attrs-child'].length).to.be.equal(4)
-    tag.unmount()
-  })
-
-  it('preserve the mount order, first the parent and then all the children', function() {
-    var correctMountingOrder = [
-        'deferred-mount',
-        'deferred-child-1',
-        'deferred-child-2',
-        'deferred-loop',
-        'deferred-loop',
-        'deferred-loop',
-        'deferred-loop',
-        'deferred-loop'
-      ],
-      mountingOrder = [],
-      cb = function(tagName, childTag) {
-        // make sure the mount event gets triggered when all the children tags
-        // are in the DOM
-        expect(document.contains(childTag.root)).to.be.equal(true)
-        mountingOrder.push(tagName)
-      },
-      tag = riot.mount('deferred-mount', { onmount: cb })[0]
-
-    expect(mountingOrder.join()).to.be.equal(correctMountingOrder.join())
-
-    tag.unmount()
-  })
-
-  it('no update should be triggered if the preventUpdate flag is set', function() {
-    var tag = riot.mount('prevent-update')[0]
-
-    expect(tag['fancy-name'].innerHTML).to.be.equal('john')
-
-    tag.root.getElementsByTagName('p')[0].onclick({})
-
-    expect(tag['fancy-name'].innerHTML).to.be.equal('john')
-
-    tag.unmount()
-  })
-
-  it('the before events get triggered', function() {
-    var eventsCount = 0,
-      tag,
-      incrementEvents = function () {
-        eventsCount++
-      }
-    riot.tag('before-events', '', function() {
-      this.on('before-mount', incrementEvents)
-      this.on('before-unmount', incrementEvents)
-    })
-    tag = riot.mount(document.createElement('before-events'))[0]
-    tag.unmount()
-    expect(eventsCount).to.be.equal(2)
-  })
-
-  it('child tags are only rendered when if-condition is truthy', function() {
-    var tag = riot.mount('if-mount')[0]
-
-    var expectL2 = function(base, exist) {
-      var ex = expect(base.tags['if-level2'])
-      exist ? ex.to.not.be.equal(undefined) : ex.to.be.equal(undefined)
-      expect($$('if-level2', base.root).length).to.be.equal(exist ? 1 : 0)
-    }
-
-    var expectCond = function(base, exist) {
-      var ex = expect(base.tags['if-level2'].tags['conditional-tag'])
-      exist ? ex.to.not.be.equal(undefined) : ex.to.be.equal(undefined)
-      expect($$('conditional-tag', base.root).length).to.be.equal(exist ? 1 : 0)
-    }
-
-    expectL2(tag.ff, false)
-    expectL2(tag.ft, false)
-
-    expectL2(tag.tf, true)
-    expectCond(tag.tf, false)
-
-    expectL2(tag.tt, true)
-    expectCond(tag.tt, true)
-
-    tag.tf.tags['if-level2'].toggleCondition()
-    expectCond(tag.tf, true)
-
-    tag.ft.toggleCondition()
-    expectL2(tag.ft, true)
-    expectCond(tag.ft, true)
-
-    tag.unmount()
-  })
-
-  it('tags under a false if statement are unmounted', function() {
-    var unmountCount = 0, cb = function() { unmountCount++ }
-    var tag = riot.mount('if-unmount', {cb: cb})[0]
-
-    // check that our child tags exist, and record their ids
-    expect(tag.tags['if-uchild'].length).to.be.equal(3)
-    var firstIds = tag.tags['if-uchild'].map(function(c) { return c._riot_id })
-
-    // set if conditions to false
-    tag.items[0].bool = false
-    tag.update({cond: false})
-
-    // ensure the tags are gone, and that their umount callbacks were triggered
-    expect(tag.tags['if-uchild']).to.be.equal(undefined)
-    expect(unmountCount).to.be.equal(3)
-
-    // set conditions back to true
-    tag.items[0].bool = true
-    tag.update({cond: true})
-
-    // ensure the tags exist, and get their ids
-    expect(tag.tags['if-uchild'].length).to.be.equal(3)
-    var secondIds = tag.tags['if-uchild'].map(function(c) { return c._riot_id })
-
-    // ensure that all of the new tags are different instances from the first time
-    var intersection = secondIds.filter(function(id2) {
-      return firstIds.indexOf(id2) > -1
-    })
-    expect(intersection.length).to.be.equal(0)
-
-    tag.unmount()
-  })
-
-  it('named refs are removed from parent when element leaves DOM', function() {
-    injectHTML('<named-unmount></named-unmount>')
-    var tag = riot.mount('named-unmount')[0]
-    tag.unmount()
-
-    expect(tag.first).to.be.equal(undefined)
-    expect(tag.second).to.be.equal(undefined)
-
-    tag.update({cond: true, items: ['third']})
-
-    expect(tag.first).to.be.an(HTMLElement)
-    expect(tag.second).to.be.an(HTMLElement)
-    expect(tag.third).to.be.an(HTMLElement)
-
-    tag.update({cond: false, items: []})
-
-    expect(tag.first).to.be.equal(undefined)
-    expect(tag.second).to.be.equal(undefined)
-    expect(tag.third).to.be.equal(undefined)
-  })
-
   it('preserve the mount order, first the parent and then all the children', function() {
     var correctMountingOrder = [
         'deferred-mount',
